add index from donatorsCab

+++ server/modules/meta/services/fileService.js
@@ -5,24 +5,92 @@ import fs from 'fs';
 import { fileSettings } from 'configuration';
 import AppError from '../../../utils/AppErrors.js';
 import { isB58, getStoragePath, checkFile, makeStorageDirs } from './helpers.js';
-import searchIndex from 'search-index';
+import { addFileIndex } from './searchService';
 
+const readFiles = (stream, multiHashRequest) => {
+  if (multiHashRequest.length==1) {
+    const filePath = checkFile(multiHashRequest[0]);
+    if (filePath) {
+      return new Promise((resolve, reject) => {
+        const preFile = fs.readFileSync(filePath, {encoding: 'utf-8'});
+        let isJSON;
+        try {
+          JSON.parse(preFile);
+          isJSON = true;
+        } catch (e) {
+          isJSON = false;
+        }
+        let file;
+        if (isJSON) {
+          console.log('json');
+          stream.writeHead(200, {'Content-Type': 'application/json'});
+          file = fs.createReadStream(filePath, {encoding: 'utf-8'});
+        } else {
+          console.log('binary');
+          stream.writeHead(200, {'X-Content-Type-Options': 'nosniff'});
+          file = fs.createReadStream(filePath);
+        }
 
-let index;
+        file
+          .on('error', (err)=> {
+            return (err.code == 'ENOENT')
+              ? reject(new AppError(406, 606))
+              : reject(new AppError(406, 605));
+          })
+          .pipe(stream);
 
-searchIndex(
-  {
-    indexPath: 'storage/index',
-    logLevel: 'error',
-  }, 
-  (err, newIndex) => {
-    if (!err) {
-      index = newIndex;
-    } else {
-      console.log(err);
+        stream
+          .on('close', () => {
+            file.destroy();
+            return reject(new AppError(406, 605));
+          })
+          .on('error', () => {
+            return reject(new AppError(406, 605));
+          })
+          .on('finish', () => {
+            stream.end();
+          });
+      });
     }
+  } else {
+    return new Promise((resolve, reject) => {
+      stream.writeHead(200, {
+        'Content-Type': 'text/plain; charset=utf-8',
+        'Cache-Control': 'no-cache',
+      });
+      const form = new FormData();
+      multiHashRequest.forEach((elem) => {
+        if (isB58(elem)) {
+          const metadataFile = checkFile(elem);
+          form.append(elem,
+            (metadataFile)
+              ? fs.createReadStream(metadataFile, {encoding: 'utf-8'})
+              : 'false'
+          );
+        } else {
+          form.append(elem, 'false');
+        }
+      });
+
+      form.on('error', (err)=> {
+        console.error(err);
+        reject(new AppError(406, 605));
+      }).pipe(stream);
+
+      stream
+        .on('close', () => {
+          form.destroy();
+          reject(new AppError(406, 605));
+        })
+        .on('error', () => {
+          reject(new AppError(406, 605));
+        })
+        .on('finish', () => {
+          stream.end();
+        });
+    });
   }
-);
+};
 
 const writeFile = (stream, tempPathFile) => {
   return new Promise((resolve, reject) => {
@@ -60,10 +128,6 @@ const writeFile = (stream, tempPathFile) => {
       .on('close', () => {
         if (!fail) {
           const tempFile = fs.readFileSync(tempPathFile);
-          const dataHashHex = crypto.createHash('sha256').update(tempFile).digest('hex');
-          const dataHashBuffer = multihash.fromHexString(dataHashHex);
-          const multiHashBuffer = multihash.encode(dataHashBuffer, 'sha2-256');
-          const multiHashB58 = multihash.toB58String(multiHashBuffer);
           let isJSON;
           try {
             const data = JSON.parse(tempFile);
@@ -72,20 +136,16 @@ const writeFile = (stream, tempPathFile) => {
           } catch (e) {
             isJSON = false;
           }
+          const dataHashHex = crypto.createHash('sha256').update(tempFile).digest('hex');
+          const dataHashBuffer = multihash.fromHexString(dataHashHex);
+          const multiHashBuffer = multihash.encode(dataHashBuffer, 'sha2-256');
+          const multiHashB58 = multihash.toB58String(multiHashBuffer);
           const metadataStoragePath = getStoragePath(multiHashB58, isJSON);
+          if (!metadataStoragePath) return localError(605);
           if (!makeStorageDirs(metadataStoragePath, isJSON)) return localError(605);
           if (!fs.existsSync(metadataStoragePath)) {
             fs.renameSync(tempPathFile, metadataStoragePath);
-            if (isJSON) {
-              fs.createReadStream(metadataStoragePath)
-                .pipe(index.feed())
-                .on('error', () => {
-                  console.log('error 1');
-                })
-                .on('finish', () => {
-                  console.log('finish 1');
-                });
-            }
+            if (isJSON) addFileIndex(metadataStoragePath);
             return resolve(multiHashB58);
           } else {
             fs.unlinkSync(tempPathFile);
@@ -96,103 +156,7 @@ const writeFile = (stream, tempPathFile) => {
   });
 };
 
-const readFiles = (stream, multiHashRequest) => {
-  if (multiHashRequest.length==1) {
-    const filePath = checkFile(multiHashRequest[0]);
-    if (filePath) {
-      return new Promise((resolve, reject) => {
-        let file;
-        if (filePath.indexOf('/json/')!=-1) {
-          console.log('json');
-          stream.writeHead(200, {'Content-Type': 'application/json'});
-          file = fs.createReadStream(filePath, {encoding: 'utf-8'});
-        } else {
-          console.log('binary');
-          stream.writeHead(200, {'X-Content-Type-Options': 'nosniff'});
-          file = fs.createReadStream(filePath);
-        }
-
-        file
-          .on('error', (err)=> {
-            return (err.code == 'ENOENT')
-              ? reject(new AppError(406, 606))
-              : reject(new AppError(406, 605));
-          })
-          .pipe(stream);
-
-        stream
-          .on('close', () => {
-            file.destroy();
-            return reject(new AppError(406, 605));
-          })
-          .on('error', () => {
-            return reject(new AppError(406, 605));
-          })
-          .on('finish', () => {
-            stream.end();
-          });
-      });
-    }
-  } else {
-    return new Promise((resolve, reject) => {
-      stream.writeHead(200, {
-        'Content-Type': 'text/plain;charset=utf-8',
-        'Cache-Control': 'no-cache',
-      });
-      const form = new FormData();
-      multiHashRequest.forEach((elem) => {
-        if (isB58(elem)) {
-          const metadataFile = checkFile(elem);
-          form.append(elem,
-            (metadataFile)
-              ? fs.createReadStream(metadataFile, {encoding: 'utf-8'})
-              : 'false'
-          );
-        } else {
-          form.append(elem, 'false');
-        }
-      });
-
-      form.on('error', (err)=> {
-        console.error(err);
-        reject(new AppError(406, 605));
-      }).pipe(stream);
-
-      stream
-        .on('close', () => {
-          form.destroy();
-          reject(new AppError(406, 605));
-        })
-        .on('error', () => {
-          reject(new AppError(406, 605));
-        })
-        .on('finish', () => {
-          stream.end();
-        });
-    });
-  }
-};
-
-const search = (text) => {
-  return new Promise((resolve, reject) => {
-    const searchResult = {};
-    index.search(text)
-      .on('data', (data) => {
-        const dataHashHex = crypto.createHash('sha256').update(JSON.stringify(data.document)).digest('hex');
-        const dataHashBuffer = multihash.fromHexString(dataHashHex);
-        const multiHashBuffer = multihash.encode(dataHashBuffer, 'sha2-256');
-        const multiHashB58 = multihash.toB58String(multiHashBuffer);
-        const metadataStoragePath = getStoragePath(multiHashB58, true);
-        if (fs.existsSync(metadataStoragePath)) searchResult[multiHashB58] = data.document;
-      })
-      .on('end', () => {
-        return resolve(searchResult);
-      });
-  });
-};
-
 export {
   writeFile,
   readFiles,
-  search,
 };
